normalize_data in MongoTestCase
