+++ src/eduid_userdb/tests/__init__.py
@@ -29,48 +29,4 @@
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 #
-import unittest
-from typing import Any, Dict, List, Tuple
-
-
-class DictTestCase(unittest.TestCase):
-    """
-    """
-    maxDiff = None
-
-    @classmethod
-    def normalize_data(cls, expected: List[Dict[str, Any]], obtained: List[Dict[str, Any]]) -> Tuple[List[Dict[str, Any]], List[Dict[str, Any]]]:
-        """
-        Remove timestamps that in general are created at different times
-        normalize the names of some attributes
-        remove attributes set to None
-        """
-        for elist in (expected, obtained):
-            for elem in elist:
-                cls.normalize_elem(elem)
-
-        return expected, obtained
-
-    @classmethod
-    def normalize_elem(cls, elem: Dict[str, Any]):
-        if 'created_ts' in elem:
-            del elem['created_ts']
-        if 'modified_ts' in elem:
-            del elem['modified_ts']
-
-        if 'application' in elem:
-            elem['created_by'] = elem.pop('application')
-
-        if 'source' in elem:
-            elem['created_by'] = elem.pop('source')
-
-        if 'credential_id' in elem:
-            elem['id'] = elem.pop('credential_id')
-
-        for key in elem:
-            if isinstance(elem[key], dict):
-                cls.normalize_elem(elem[key])
-
-        for key in ('created_by', 'application', 'verified_ts', 'verified_by'):
-            if key in elem and elem[key] is None:
-                del elem[key]
+from eduid_userdb.testing import DictTestCase
